requirements.txt
streamlit
langchain
openai
python-dotenv
pandas
langchain-experimental
langchain-groq
langchain-community


matplotlib


seaborn


sqlalchemy

